[feat] 리덕스 스토어 세팅
