+++ src/store/index.ts
@@ -1,0 +1,21 @@
+import {configureStore} from '@reduxjs/toolkit';
+import {useDispatch} from 'react-redux';
+import rootReducer from './reducer';
+
+const store = configureStore({
+  reducer: rootReducer,
+
+  //flipper와 연결하기 위해 middleware 사용.
+  middleware: getDefaultMiddleware => {
+    if (__DEV__) {
+      const createDebugger = require('redux-flipper').default;
+      return getDefaultMiddleware().concat(createDebugger());
+    }
+    return getDefaultMiddleware();
+  },
+});
+export default store;
+
+//typeScript때문에 작성.
+export type AppDispatch = typeof store.dispatch;
+export const useAppDispatch = () => useDispatch<AppDispatch>();
