+++ src/store/reducer.ts
@@ -1,7 +1,9 @@
 import {combineReducers} from 'redux';
+import storeMapSlice from '../slices/storeMap';
 
 const rootReducer = combineReducers({
   //slice 내용 적으면 됨.
+  storeMap: storeMapSlice.reducer,
 });
 
 //typeScript의 Type error 피하기 위해 작성함.
